Move options and functions to AppInitializer, Add WindowGrouping type
