+++ test/demo/utils/AppInitializer.ts
@@ -7,13 +7,11 @@ import {delay} from '../../provider/utils/delay';
 import {dragSideToSide} from '../../provider/utils/dragWindowTo';
 import {tabWindowsTogether} from '../../provider/utils/tabWindowsTogether';
 
-// import { randomCoordinate } from '../workspaces/basicSaveAndRestore.test';
-
 interface ParamBase {
     childWindows: fin.WindowOptions[];
 }
 
-interface ProgrammaticParam extends ParamBase {
+interface ProgrammaticParams extends ParamBase {
     createType: 'programmatic';
     appOptions: fin.ApplicationOptions;
 }
@@ -23,7 +21,11 @@ interface ManifestParams extends ParamBase {
     manifestUrl: string;
 }
 
-export type AppInitializerInfo = ProgrammaticParam|ManifestParams;
+export type AppInitializerInfo = ProgrammaticParams|ManifestParams;
+
+// A WindowGrouping is an array of array of numbers that corresponds to two windows grouped. 
+// e.g. [[0, 1], [2, 3]] would mean that out of an array of 4 windows, win0 and win1 should be grouped, and win2 and win3 should be grouped.
+export type WindowGrouping = number[][];
 
 export interface TestApp {
     uuid: string;
@@ -44,12 +46,75 @@ const CHILD_WINDOW_BASE = {
     name: 'BASE'
 };
 
+export const OPTIONS_BASE = {
+    uuid: 'BASE',
+    url: 'http://localhost:1337/test/registeredApp.html',
+    name: 'BASE',
+    autoShow: true,
+    saveWindowState: false,
+    defaultTop: 100,
+    defaultLeft: 100,
+    defaultHeight: 250,
+    defaultWidth: 250
+};
+
+export const APP_INITIALIZER_BASE = {
+    appOptions: OPTIONS_BASE,
+    createType: 'programmatic',
+    childWindows: []
+};
+
 function childYCoordinate(appNum: number, childNum: number) {
     return (appNum * 275) + (childNum * 50) + 100;
 }
 
 function childXCoordinate(appNum: number, childNum: number) {
     return ((appNum + childNum) * 280) + 300;
+}
+
+export function appYCoordinate(appTitleNumber: number) {
+    return ((appTitleNumber - 1) * 275) + 50;
+}
+
+let appTitleNumber = 0;
+
+export function createAppsArray(numAppsToCreate: number, numberOfChildren: number) {
+    const appsArray = [];
+    while (numAppsToCreate > 0) {
+        // Set the app information
+        appTitleNumber++;
+        const id = 'test-app' + appTitleNumber;
+
+        // Set the child window information
+        const childWindows = [];
+        let childTitleNumber = numberOfChildren;
+        while (childTitleNumber > 0) {
+            childWindows.push({name: `test${appTitleNumber}-child${childTitleNumber}`});
+            childTitleNumber--;
+        }
+
+        // Save the app information
+        const defaultTop = appYCoordinate(numAppsToCreate);
+        const appOptions = {...OPTIONS_BASE, uuid: id, name: id, defaultTop, defaultLeft: 100};
+        appsArray.push({...APP_INITIALIZER_BASE, appOptions, childWindows});
+
+        numAppsToCreate--;
+    }
+
+    return appsArray;
+}
+
+export function createWindowGroupings(numApps: number, children: number) {
+    const totalWindows = (numApps) + (numApps * children);
+
+    switch (totalWindows) {
+        case 2:
+            return [[[0, 1]]];
+        case 4:
+            return [[[0, 1], [2, 3]], [[0, 2], [1, 3]], [[0, 3], [1, 2]]];
+        default:
+            return [];
+    }
 }
 
 export class AppInitializer {
@@ -99,7 +164,7 @@ export class AppInitializer {
         return result;
     }
 
-    public async snapWindows(snapWindowGrouping: number[][], windows: _Window[]): Promise<void> {
+    public async snapWindows(snapWindowGrouping: WindowGrouping, windows: _Window[]): Promise<void> {
         for (const group of snapWindowGrouping) {
             const win1 = windows[group[0]];
             const win2 = windows[group[1]];
@@ -107,7 +172,7 @@ export class AppInitializer {
         }
     }
 
-    public async tabWindows(tabWindowGrouping: number[][], windows: _Window[]): Promise<void> {
+    public async tabWindows(tabWindowGrouping: WindowGrouping, windows: _Window[]): Promise<void> {
         for (const group of tabWindowGrouping) {
             const win1 = windows[group[0]];
             const win2 = windows[group[1]];
